receiver/requirements.txt
Flask==2.3.3
google-cloud-pubsub==2.20.0
gunicorn==21.2.0